fix(ingest): clinical_trials — add conditions/interventions fields, link-table inserts, fix api_page_size default
- Change api_page_size type from u32 to usize, default from 1000 to 100
- Add conditions/interventions Vec<String> fields to AactStudyRow (initialized empty from CSV parser and API fetcher)
- Update api_fetcher page_size param to usize to match config type
- Add trial_disease_links and trial_intervention_links inserts in storage after clinical_trials upsert

+++ crates/bdp-ingest/src/pipelines/clinical_trials/api_fetcher.rs
@@ -17,7 +17,7 @@ pub async fn fetch_updated_studies(
     client: &Client,
     base_url: &str,
     from_date: NaiveDate,
-    page_size: u32,
+    page_size: usize,
     max_retries: u32,
 ) -> Result<Vec<serde_json::Value>> {
     let date_str = from_date.format("%Y-%m-%d").to_string();
